autonomous-research-agent/requirements.txt
chromadb==1.5.7
fastapi==0.135.3
langchain-core==1.2.28
langchain-openai==1.1.12
langgraph==1.1.6
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.9
langgraph-sdk==0.3.13
mcp==1.27.0
openai==2.31.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
python-dotenv==1.2.2
tavily-python==0.7.23
uvicorn==0.44.0
